+++ angular/lib/src/core/metadata.dart
@@ -80,105 +80,6 @@ class Directive {
   /// element.
   final String selector;
 
-  /// Events, actions, properties, and attributes related to the host element.
-  ///
-  /// ## Host listeners
-  /// Specifies which DOM events the directive listens to via a set of
-  /// '(_event_)' to _statement_ key-value pairs:
-  ///
-  /// - _event_: the DOM event that the directive listens to
-  /// - _statement_: the statement to execute when the event occurs
-  ///
-  /// If the evaluation of the statement returns [false], then [preventDefault]
-  /// is applied on the DOM event.
-  ///
-  /// To listen to global events, a target must be added to the event name.
-  /// The target can be `window`, `document`, or `body`.
-  ///
-  /// When writing a directive event binding, you can also refer to the `$event`
-  /// local variable.
-  ///
-  /// The following example declares a directive that attaches a click listener
-  /// to the button and counts clicks.
-  ///
-  /// ```dart
-  /// @Directive(
-  ///   selector: 'button[counting]',
-  ///   host: const {
-  ///     '(click)': 'onClick($event.target)'
-  ///   })
-  /// class CountClicks {
-  ///   var numberOfClicks = 0;
-  ///
-  ///   void onClick(btn) {
-  ///     print("Button $btn, number of clicks: ${numberOfClicks++}.");
-  ///   }
-  /// }
-  ///
-  /// @Component(
-  ///   selector: 'app',
-  ///   template: `<button counting>Increment</button>`,
-  ///   directives: const [CountClicks])
-  /// class App {}
-  /// ```
-  ///
-  /// ## Host property bindings
-  /// Specifies which DOM properties the directive updates.
-  ///
-  /// Angular automatically checks host property bindings during change
-  /// detection. If a binding changes, it will update the host element of the
-  /// directive.
-  ///
-  /// The following example creates a directive that sets the `valid` and
-  /// `invalid` classes on the DOM element that has ngModel directive on it.
-  ///
-  ///     @Directive(
-  ///       selector: '[ngModel]',
-  ///       host: {
-  ///         '[class.valid]': 'valid',
-  ///         '[class.invalid]': 'invalid'
-  ///       }
-  ///     )
-  ///     class NgModelStatus {
-  ///       NgModel control;
-  ///
-  ///       NgModelStatus(this.control);
-  ///       get valid => control.valid;
-  ///       get invalid => control.invalid;
-  ///     }
-  ///
-  ///     @Component({
-  ///       selector: 'app',
-  ///       template: `<input [(ngModel)]="prop">`,
-  ///       directives: [formDirectives, NgModelStatus]
-  ///     })
-  ///     class App {
-  ///       prop;
-  ///     }
-  ///
-  ///     bootstrap(App);
-  ///
-  /// ## Attributes
-  ///
-  /// Specifies static attributes that should be propagated to a host element.
-  ///
-  /// ### Example
-  ///
-  /// In this example using `my-button` directive (ex.: `<div my-button></div>`)
-  /// on a host element (here: `<div>` ) will ensure that this element will get
-  /// the "button" role.
-  ///
-  /// ```dart
-  /// @Directive(
-  ///   selector: '[my-button]',
-  ///   host: const {
-  ///     'role': 'button'
-  ///   })
-  /// class MyButton {}
-  /// ```
-  @Deprecated('Use @HostBinding() on a getter or @HostListener on a method')
-  final Map<String, String> host;
-
   /// The set of injectable objects that are visible to the directive and
   /// its light DOM children.
   ///
@@ -229,8 +130,6 @@ class Directive {
 
   const Directive({
     @required this.selector,
-    @Deprecated('Use @HostBinding() on a getter or @HostListener on a method')
-        this.host,
     this.providers,
     this.exportAs,
     this.visibility: Visibility.local,
@@ -240,8 +139,7 @@ class Directive {
 /// Declare reusable UI building blocks for an application.
 ///
 /// Each Angular component requires a single `@Component` annotation. The
-/// `@Component` annotation specifies when a component is instantiated, and
-/// which properties and hostListeners it binds to.
+/// `@Component` annotation specifies when a component is instantiated.
 ///
 /// When a component is instantiated, Angular
 ///
@@ -338,8 +236,6 @@ class Component extends Directive {
 
   const Component({
     String selector,
-    @Deprecated('Use @HostBinding() on a getter or @HostListener on a method')
-        Map<String, String> host,
     String exportAs,
     List providers,
     Visibility visibility: Visibility.local,
@@ -356,7 +252,6 @@ class Component extends Directive {
     this.encapsulation,
   }) : super(
           selector: selector,
-          host: host,
           exportAs: exportAs,
           providers: providers,
           visibility: visibility,