+++ angular/CHANGELOG.md
@@ -3,6 +3,10 @@
 *   A directive with `Visibility.local` may now be injected by another directive
     on the same host element, or by a descendant within the same view.
 
+*   Removed support for (deprecated) `host: const { ... }` syntax in a
+    `@Directive` or `@Component` annotation. This can be easily migrated to use
+    `@HostBinding` or `@HostListener`.
+
 ## 5.0.0-alpha+13
 
 * Pins `angular_ast` and `angular_compiler` to avoid future versioning issues.